+++ lib/core/error/failure.dart
@@ -3,6 +3,12 @@ import 'package:equatable/equatable.dart';
 abstract class Failure extends Equatable {}
 
 class ServerFailure extends Failure {
+  final String failureMessage;
+
+  ServerFailure({
+    this.failureMessage = 'Server failure',
+  });
+  
   @override
-  List<Object> get props => null;
+  List<Object> get props => [failureMessage];
 }
