+++ lib/core/usecase/usecase.dart
@@ -1,7 +1,12 @@
 import 'package:cardiompp/core/error/failure.dart';
 import 'package:dartz/dartz.dart';
+import 'package:equatable/equatable.dart';
 
 abstract class UseCase<Type, Params> {
-  Future<bool> call(Params params);
-  // Future<Either<Failure, bool>> call(Params params);
+  Future<Either<Failure, Type>> call(Params params);
+}
+
+class NoParams extends Equatable {
+  @override
+  List<Object> get props => [];
 }
