+++ lib/features/login/data/datasources/local_datasource.dart
@@ -52,8 +52,8 @@ class LocalDataSourceImpl implements LocalDataSource {
     if (preferences.containsKey(key)) {
       var tokenTime = preferences.getInt(key);
 
-      print('tokenTime Armazenado: $tokenTime');
-      print('tokenTime Atual: $timeNow');
+      // print('tokenTime Armazenado: $tokenTime');
+      // print('tokenTime Atual: $timeNow');
 
       if (timeNow - tokenTime > 29*60*1000) {
         return true;
