+++ lib/features/login/domain/entities/authenticate.dart
