+++ lib/features/login/domain/entities/loginApp.dart
