+++ lib/main.dart
@@ -1,49 +1,26 @@
-import 'package:cardiompp/features/presentation/pages/loged_page.dart';
+import 'package:cardiompp/features/login/presentation/pages/loged_page.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'features/presentation/blocs/login_bloc/login_bloc.dart';
-import 'features/presentation/pages/login_page.dart';
+import 'features/login/presentation/bloc/login_bloc.dart';
+import 'features/login/presentation/pages/login_page.dart';
 
 void main() => runApp(MyApp());
 
 class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(
-      title: 'Login Challenge',
-      routes: <String, WidgetBuilder>{
-        LogedPage.routeName: (context) =>
-          LogedPage(title: 'Loged in Mobile TCC'),
-        LoginPage.routeName: (context) => LoginPage(),
-
+    return BlocProvider<LoginBloc> (
+      create: (BuildContext context) {
+        return LoginBloc();
       },
-      debugShowCheckedModeBanner: false,
-      theme: ThemeData(
-        primarySwatch: Colors.deepPurple,
+      child: MaterialApp(
+        home: LoginPage(),
+        debugShowCheckedModeBanner: false,
+        routes: {
+          '/login': (context) => LoginPage(),
+          '/logged': (context) => LogedPage(),
+        },
       ),
-      home: MultiBlocProvider(
-        providers: [
-          BlocProvider<LoginBloc>(
-            create: (context) => LoginBloc(),
-          ),
-        ],
-        child: BlocBuilder<LoginBloc, LoginState>(
-          builder: (context, state) {
-            if(state is LoginStartState) {
-              return Center(
-                child: CircularProgressIndicator(),
-              );
-            }
-            if(state is LoginErrorState) {
-              return LoginPage(errorMessage: state.message);
-            }
-            if(state is LoginSuccessState) {
-              return LogedPage(doctor: state.doctorState);
-            }
-            return LoginPage();
-          },  
-        ),
-      )
     );
   }
 }
