+++ pubspec.yaml
@@ -25,16 +25,11 @@ dependencies:
   equatable: ^1.1.1
   dartz: ^0.9.0
   connectivity: ^0.4.8+2
-  data_connection_checker: ^0.3.4
   http: ^0.12.1
   shared_preferences: ^0.5.7
-  flutter_cache_manager: ^1.2.2
   path_provider: ^1.6.7
   dio: ^3.0.9
   bloc: ^4.0.0
-  cookie_jar: ^1.0.1
-  rxdart: ^0.24.0
-  client_cookie: ^2.2.1
   
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
